# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: EXPP API
  version: "1.0.0"
  description: >
    EXPP — Educational Workflow Platform v1.
    All endpoints require authentication via the `expp_session` HTTP-only cookie
    unless noted as public.
  contact:
    name: EXPP Project
servers:
  - url: /api
    description: Next.js App Router API
...
components:
  schemas:
    ApiSuccess:
      type: object
      required: [success, data, traceId]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          description: Response payload (schema varies per endpoint)
        traceId:
          type: string
    ApiError:
      type: object
      required: [success, error, code, traceId]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string
          enum: [AUTH_ERROR, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, RATE_LIMITED]
        traceId:
          type: string
    Role:
      type: string
      enum: [STUDENT, TEACHER, ADMIN]
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: expp_session
security:
  - sessionCookie: []

# ── Paths ──────────────────────────────────────────────────────────────────
paths:
...
  /ai-help:
    post:
      tags: [ai-help]
      summary: Send AI help request (policy-gated)
      description: |
        STUDENT role required. Three-layer policy enforcement:
        1. UI layer (client-side mode check)
        2. Server layer (policy-engine.ts)
        3. Gateway layer (ai-help-gateway.ts)

        Blocked requests return 403 and are logged in the audit trail.
        All requests (allowed and blocked) are logged as HelpRequest records.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attemptId, message]
              properties:
                attemptId: { type: string }
                message: { type: string }
      responses:
        "200":
          description: AI response returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "403":
          description: Help blocked by policy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }

  # ── Assessment ────────────────────────────────────────────────────────────
  /assessment:
    get:
      tags: [assessment]
      summary: List submitted attempts for a distribution
      description: TEACHER role required. Query param distributionId required.
      parameters:
        - name: distributionId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Submission list with assessment status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
  /assessment/{attemptId}:
    get:
      tags: [assessment]
      summary: Get assessment detail for an attempt
      description: TEACHER role required. Returns assessment status, autoCheckStatus, latestAiRun, and aiRecommendation when available.
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Assessment with auto-check results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }

    post:
      tags: [assessment]
      summary: Get or create assessment for an attempt
      description: TEACHER role required. Ensures baseline deterministic auto-check exists.
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Assessment baseline ready
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
...
  /assessment/{attemptId}/review:
    patch:
      tags: [assessment]
      summary: Save teacher review (grade + comment)
      description: TEACHER role required. Transitions to REVIEWED. Blocked if already PUBLISHED.
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manualGrade, maxGrade]
              properties:
                manualGrade: { type: number, minimum: 0 }
                maxGrade: { type: number, minimum: 0 }
                comment: { type: string, nullable: true }
                itemOverrides:
                  type: array
                  items:
                    type: object
                    properties:
                      itemOrder: { type: integer, minimum: 1 }
                      teacherScore: { type: number, minimum: 0 }
                      reason: { type: string }
      responses:
        "200":
          description: Assessment updated to REVIEWED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "422":
          description: Grade > maxGrade or assessment already PUBLISHED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /assessment/{attemptId}/publish:
    post:
      tags: [assessment]
      summary: Publish assessment result (REVIEWED → PUBLISHED)
      description: TEACHER role required. Allowed only from REVIEWED status. Result becomes visible to student and immutable after publication.
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "422":
          description: Not REVIEWED or already PUBLISHED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /admin/users/{id}:
    patch:
      tags: [admin]
      summary: Update user role or active status
      description: ADMIN role required. Cannot modify your own account. Cannot demote the last admin.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isActive: { type: boolean }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "422":
          description: Self-modification or last-admin guard
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /admin/audit:
    get:
      tags: [admin]
      summary: Audit log with filters
      description: |
        ADMIN: full log with optional userId filter.
        TEACHER: scoped to own events only.
      parameters:
        - name: action
          in: query
          schema: { type: string }
          description: e.g. "auth.login", "attempt.submitted"
        - name: userId
          in: query
          schema: { type: string }
          description: Admin only — filter by user
        - name: entityType
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
      responses:
        "200":
          description: Paginated audit log
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }